+++ src/qcoder/__init__.py
@@ -1,3 +1,3 @@
 __all__ = []
-__version__ = "0.6.0a14"
+__version__ = "0.6.0a15"
 file = __file__